api_server/requirements.txt
fastapi[standard]
uvicorn==0.32.0
pydantic
langchain
langchain-core
langchain-google-genai
langchain-openai
langchain-community
langgraph
langgraph-checkpoint-postgres


pgvector
psycopg[binary,pool]
psycopg2-binary
pyjwt
passlib
sqlalchemy
aiosmtplib
cryptography
fuzzywuzzy
langchain-groq
bcrypt
certifi